Update to 1.20.1

+++ src/main/java/dev/nikomaru/horseinfonoticer/utils/RenderUtil.java
@@ -48,7 +48,7 @@ public class RenderUtil {
         for (var s : infoString) {
             width = Math.max(mc.textRenderer.getWidth(s), width);
         }
-        var widthHalf = width / 2.0F;
+        var widthHalf = width / 2;
 
         var matrix4f = matrixStackIn.peek().getPositionMatrix();
         var f1 = mc.options.getTextBackgroundOpacity(0.4f);
@@ -58,7 +58,7 @@ public class RenderUtil {
         var j = ((int) (f1 * 255.0F) << 24) + ((int) (r * 255.0F) << 16) + ((int) (g * 255.0F) << 8) + ((int) (b * 255.0F));
 
         for (var i = 0; i < infoString.size(); i++) {
-            mc.textRenderer.draw(infoString.get(i), -widthHalf, baseY + fontHeight * i, (i == 0) ? titleColor.getRGB() :
+            mc.textRenderer.draw(infoString.get(i), -widthHalf, (int) baseY + fontHeight * i, (i == 0) ? titleColor.getRGB() :
                             fontColor.getRGB(),
                     false, matrix4f, bufferIn, TextRenderer.TextLayerType.NORMAL, j, packedLightIn);
         }
